Move some classes
